requirements.txt
altair==6.0.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
ImageIO==2.37.3
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-loader==0.5
MarkupSafe==3.0.3
narwhals==2.19.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.1
pandas==3.0.2
pillow==12.2.0
protobuf==7.34.1
pyarrow==23.0.1
pydeck==0.9.2
python-dateutil==2.9.0.post0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
smmap==5.0.3
streamlit==1.56.0
tenacity==9.1.4
threadpoolctl==3.6.0
tifffile==2026.4.11
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
urllib3==2.6.3
